Adds a package.json to make project installable by npm.
This is kind of just for my own use at the moment, so npm can install pagedown-extra via dependencies using the my github repo URL. It's not intended to be published on the npm registry, or at least not without talking with the original author.

Both myself (ncraike) and jmcmanus are listed as contributors in the package.json. As of this moment, jmcmanus is the primary author; I'm just making some minor changes to make pagedown-extra usable in node.js.